% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotools.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Calculates distance between two point using google API
You have to create own API key here: https://developers.google.com/maps/documentation/distance-matrix/}
\usage{
calculateDistance(origin, target, mode = "driving", apikey)
}
\arguments{
\item{origin}{start of journer}

\item{target}{eng of the journey}

\item{mode}{one of 'driving' (default), walking, bycycling or transit (public transport)}

\item{apikey}{}
}
\value{
list with distance in kilometers and time in minutes
}
\description{
Calculates distance between two point using google API
You have to create own API key here: https://developers.google.com/maps/documentation/distance-matrix/
}
