% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotools.R
\name{freegeoip}
\alias{freegeoip}
\title{Use freegeoip.net for converting IP address to geo location}
\usage{
freegeoip(ip, format = ifelse(length(ip) == 1, "list", "dataframe"))
}
\arguments{
\item{ip}{vector strings with IP address}

\item{format}{'list' or 'dataframe' default value is list for ip vector with only one value}
}
\description{
Use freegeoip.net for converting IP address to geo location
}
