% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceIndexers.R
\name{heurekaFeed2df}
\alias{heurekaFeed2df}
\title{Customizable XML parser for product feeds.}
\usage{
heurekaFeed2df(doc, xpath = "//SHOPITEM", isXML = TRUE, usewhich = TRUE,
  verbose = TRUE, skipTags = c("PARAM", "DELIVERY"))
}
\arguments{
\item{doc}{XML parsed document or string or file}

\item{xpath}{the xpath to shopitem in the feed}

\item{isXML}{indicated if doc is parsed XML.}

\item{usewhich}{is experimental. I looks like function works faster when set to true}

\item{verbose}{sets debug messages on}

\item{skipTags}{is a collection of tags which will not be processed. Typically multiplve values in PARAM and DELIVERY}
}
\description{
\code{heurekaFeed2df(doc)} create new datafraame containing the flat structure of the Heureka product feed. It skips problematics tags.
}
\details{
This  is a XML parser optimised for Heureka feed. It can replaces \code{XML::xmlTodataframe} in all cases when this function fails for some reason.
}
\examples{

library(XML)
doc <- xmlParse("http://my.eshop.cz/heureka.xml")
df <- heurekaFeed2df(doc)

}

